requirements.txt
pyinstaller==4.3
black
pywebview
requests
